requirements.txt
alpaca-py>=0.8.0
pandas>=2.1.0
numpy>=1.26.0
yfinance>=0.2.36
requests>=2.31.0
python-dotenv>=1.0.0 